# app/controllers/users_controller.go
package controllers

import (
	"context"
	requests "users-service/app/models/requests"
	"users-service/app/models/responses"
	"users-service/app/services"
	"users-service/app/utils"

	"github.com/gin-gonic/gin"
	"github.com/rs/zerolog"
)

type UsersController struct {
	UsersService services.UsersServiceInterface
	Logger       zerolog.Logger
}

type UsersControllerInterface interface {
	RespondFailure(ctx *gin.Context, statusCode int, errMsg string)
	RespondSuccess(ctx *gin.Context, statusCode int, data interface{})

	UpdateUser(ctx *gin.Context)
	UpdateDetails(ctx *gin.Context)
	UpdateSettings(ctx *gin.Context)
	RemoveAccount(ctx *gin.Context)
}

var _ UsersControllerInterface = (*UsersController)(nil)
var _ StandardController = (*UsersController)(nil)

func NewUsersController(
	usersService services.UsersServiceInterface,
	logger zerolog.Logger) *UsersController {

	return &UsersController{
		UsersService: usersService,
		Logger:       logger,
	}
}

func (c *UsersController) Log() *zerolog.Logger {
	return &c.Logger
}

func (c *UsersController) RespondFailure(ctx *gin.Context, statusCode int, errMsg string) {
	requestID := utils.GetRequestID(ctx)

	ctx.JSON(statusCode, responses.ErrorResponse{
		Status:    "failure",
		Error:     errMsg,
		RequestID: requestID,
	})
}

func (c *UsersController) RespondSuccess(ctx *gin.Context, statusCode int, data interface{}) {
	requestID := utils.GetRequestID(ctx)

	ctx.JSON(statusCode, responses.SuccessResponse{
		Status:    "success",
		Data:      data,
		RequestID: requestID,
	})
}

// UpdateUser godoc
// @Summary      Update user's basic information
// @Description  Updates the authenticated user's basic profile information (e.g., name, email).
// @Tags         users
// @Accept       json
// @Produce      json
// @Param        Authorization header    string            true  "Bearer token"
// @Param        body      body      requests.User  true  "User data to update"
// @Success      200       {object}  responses.SuccessResponse{data=responses.User}  "OK - User updated successfully"
// @Failure      400       {object}  responses.ErrorResponse "Bad Request - Invalid input data"
// @Failure      401       {object}  responses.ErrorResponse "Unauthorized - Invalid or missing token"
// @Failure      422       {object}  responses.ErrorResponse "Unprocessable Entity - Update failed"
// @Router       /api/v1/users/update [patch]
func (c *UsersController) UpdateUser(ctx *gin.Context) {
	handleAuthenticatedRequest(ctx, c, func(reqCtx context.Context, userID int, req *requests.User) (*responses.User, error) {
		user, err := c.UsersService.UpdateUser(reqCtx, userID, req)
		if err != nil {
			return nil, err
		}
		return &responses.User{User: user}, nil
	})
}

// UpdateDetails godoc
// @Summary      Update user's details
// @Description  Updates the authenticated user's additional details.
// @Tags         users
// @Accept       json
// @Produce      json
// @Param        Authorization header    string            true  "Bearer token"
// @Param        body      body      requests.Details  true  "User details to update"
// @Success      200       {object}  responses.SuccessResponse{data=responses.User}  "OK - User details updated successfully"
// @Failure      400       {object}  responses.ErrorResponse "Bad Request - Invalid input data"
// @Failure      401       {object}  responses.ErrorResponse "Unauthorized - Invalid or missing token"
// @Failure      422       {object}  responses.ErrorResponse "Unprocessable Entity - Update failed"
// @Router       /api/v1/users/details/update [patch]
func (c *UsersController) UpdateDetails(ctx *gin.Context) {
	handleAuthenticatedRequest(ctx, c, func(reqCtx context.Context, userID int, req *requests.Details) (*responses.User, error) {
		user, err := c.UsersService.UpdateDetails(reqCtx, userID, req)
		if err != nil {
			return nil, err
		}
		return &responses.User{User: user}, nil
	})
}

// UpdateSettings godoc
// @Summary      Update user's settings
// @Description  Updates the authenticated user's application settings (e.g., 2FA, night mode).
// @Tags         users
// @Accept       json
// @Produce      json
// @Param        Authorization header    string            true  "Bearer token"
// @Param        body      body      requests.Settings  true  "User settings to update"
// @Success      200       {object}  responses.SuccessResponse{data=responses.User}  "OK - User settings updated successfully"
// @Failure      400       {object}  responses.ErrorResponse "Bad Request - Invalid input data"
// @Failure      401       {object}  responses.ErrorResponse "Unauthorized - Invalid or missing token"
// @Failure      422       {object}  responses.ErrorResponse "Unprocessable Entity - Update failed"
// @Router       /api/v1/users/settings/update [patch]
func (c *UsersController) UpdateSettings(ctx *gin.Context) {
	handleAuthenticatedRequest(ctx, c, func(reqCtx context.Context, userID int, req *requests.Settings) (*responses.User, error) {
		user, err := c.UsersService.UpdateSettings(reqCtx, userID, req)
		if err != nil {
			return nil, err
		}
		return &responses.User{User: user}, nil
	})
}

// RemoveAccount godoc
// @Summary      Remove user account
// @Description  Permanently removes the authenticated user's account. This action is irreversible.
// @Tags         users
// @Accept       json
// @Produce      json
// @Param        Authorization header    string            true  "Bearer token"
// @Success      200       {object}  responses.SuccessResponse{data=string}  "OK - Account removed successfully"
// @Failure      401       {object}  responses.ErrorResponse "Unauthorized - Invalid or missing token"
// @Failure      422       {object}  responses.ErrorResponse "Unprocessable Entity - Removal failed"
// @Router       /api/v1/users/remove [delete]
func (c *UsersController) RemoveAccount(ctx *gin.Context) {
	handleAuthenticatedRequest(ctx, c, func(reqCtx context.Context, userID int, req *requests.Empty) (string, error) {
		err := c.UsersService.RemoveAccount(reqCtx, userID)
		if err != nil {
			return "failed to remove user", err
		}
		return "successfully removed user", nil
	})
}