# docker/docs/swagger.yaml
definitions:
  users-service_app_models_ent.ActivationCode:
    properties:
      code:
        description: Code holds the value of the "code" field.
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.ActivationCodeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ActivationCodeQuery when eager-loading is set.
      expires_at:
        description: ExpiresAt holds the value of the "expires_at" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      used:
        description: Used holds the value of the "used" field.
        type: boolean
    type: object
  users-service_app_models_ent.ActivationCodeEdges:
    properties:
      owner:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.User'
        description: Owner holds the value of the owner edge.
    type: object
  users-service_app_models_ent.ResetCode:
    properties:
      code:
        description: Code holds the value of the "code" field.
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.ResetCodeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ResetCodeQuery when eager-loading is set.
      expires_at:
        description: ExpiresAt holds the value of the "expires_at" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
    type: object
  users-service_app_models_ent.ResetCodeEdges:
    properties:
      owner:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.User'
        description: Owner holds the value of the owner edge.
    type: object
  users-service_app_models_ent.RolePermission:
    properties:
      add:
        description: Add holds the value of the "add" field.
        type: boolean
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      delete:
        description: Delete holds the value of the "delete" field.
        type: boolean
      edges:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.RolePermissionEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the RolePermissionQuery when eager-loading is set.
      edit:
        description: Edit holds the value of the "edit" field.
        type: boolean
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      view:
        description: View holds the value of the "view" field.
        type: boolean
    type: object
  users-service_app_models_ent.RolePermissionEdges:
    properties:
      user_role:
        description: UserRole holds the value of the user_role edge.
        items:
          $ref: '#/definitions/users-service_app_models_ent.UserRole'
        type: array
    type: object
  users-service_app_models_ent.SecondFactorCode:
    properties:
      code:
        description: Code holds the value of the "code" field.
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.SecondFactorCodeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the SecondFactorCodeQuery when eager-loading is set.
      expires_at:
        description: ExpiresAt holds the value of the "expires_at" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      used:
        description: Used holds the value of the "used" field.
        type: boolean
    type: object
  users-service_app_models_ent.SecondFactorCodeEdges:
    properties:
      owner:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.User'
        description: Owner holds the value of the owner edge.
      target_user_device:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.UserDevice'
        description: TargetUserDevice holds the value of the target_user_device edge.
    type: object
  users-service_app_models_ent.User:
    properties:
      active:
        description: Active holds the value of the "active" field.
        type: boolean
      blacklisted:
        description: Blacklisted holds the value of the "blacklisted" field.
        type: boolean
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.UserEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      mail:
        description: Mail holds the value of the "mail" field.
        type: string
      password:
        description: Password holds the value of the "password" field.
        type: string
      phone:
        description: Phone holds the value of the "phone" field.
        type: string
      removed:
        description: Removed holds the value of the "removed" field.
        type: boolean
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      verified:
        description: Verified holds the value of the "verified" field.
        type: boolean
    type: object
  users-service_app_models_ent.UserAction:
    properties:
      action:
        description: Action holds the value of the "action" field.
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      details:
        description: Details holds the value of the "details" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.UserActionEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserActionQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      type:
        description: Type holds the value of the "type" field.
        type: string
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
    type: object
  users-service_app_models_ent.UserActionEdges:
    properties:
      author:
        description: Author holds the value of the author edge.
        items:
          $ref: '#/definitions/users-service_app_models_ent.User'
        type: array
    type: object
  users-service_app_models_ent.UserDetails:
    properties:
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.UserDetailsEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserDetailsQuery when eager-loading is set.
      first_name:
        description: FirstName holds the value of the "first_name" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      last_name:
        description: LastName holds the value of the "last_name" field.
        type: string
      name:
        description: Name holds the value of the "name" field.
        type: string
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
    type: object
  users-service_app_models_ent.UserDetailsEdges:
    properties:
      owner:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.User'
        description: Owner holds the value of the owner edge.
    type: object
  users-service_app_models_ent.UserDevice:
    properties:
      browser_name:
        description: Browser name (e.g., Chrome, Firefox).
        type: string
      browser_version:
        description: Browser version (e.g., 126.0).
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.UserDeviceEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserDeviceQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      ip_address:
        description: The last known IP address of the device.
        type: string
      last_seen_at:
        description: Timestamp of the last time this device was used for authentication.
        type: string
      name:
        description: Name of the device
        type: string
      os_name:
        description: Operating system name (e.g., Windows, macOS).
        type: string
      os_version:
        description: Operating system version (e.g., 11, 14.5).
        type: string
      token:
        description: The unique fingerprint visitorId from FingerprintJS.
        type: string
      type:
        description: type of the device
        type: string
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      user_agent:
        description: The full User-Agent string from the last request.
        type: string
    type: object
  users-service_app_models_ent.UserDeviceEdges:
    properties:
      owner:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.User'
        description: Owner holds the value of the owner edge.
      second_factor_codes:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.SecondFactorCode'
        description: SecondFactorCodes holds the value of the second_factor_codes
          edge.
      user_settings_2fa:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.UserSettings'
        description: UserSettings2fa holds the value of the user_settings_2fa edge.
    type: object
  users-service_app_models_ent.UserEdges:
    properties:
      activation_code:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.ActivationCode'
        description: ActivationCode holds the value of the activation_code edge.
      reset_code:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.ResetCode'
        description: ResetCode holds the value of the reset_code edge.
      second_factor_code:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.SecondFactorCode'
        description: SecondFactorCode holds the value of the second_factor_code edge.
      user_actions:
        description: UserActions holds the value of the user_actions edge.
        items:
          $ref: '#/definitions/users-service_app_models_ent.UserAction'
        type: array
      user_details:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.UserDetails'
        description: UserDetails holds the value of the user_details edge.
      user_devices:
        description: UserDevices holds the value of the user_devices edge.
        items:
          $ref: '#/definitions/users-service_app_models_ent.UserDevice'
        type: array
      user_roles:
        description: UserRoles holds the value of the user_roles edge.
        items:
          $ref: '#/definitions/users-service_app_models_ent.UserRole'
        type: array
      user_settings:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.UserSettings'
        description: UserSettings holds the value of the user_settings edge.
      verification_code:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.VerificationCode'
        description: VerificationCode holds the value of the verification_code edge.
    type: object
  users-service_app_models_ent.UserRole:
    properties:
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      description:
        description: Description holds the value of the "description" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.UserRoleEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserRoleQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
    type: object
  users-service_app_models_ent.UserRoleEdges:
    properties:
      owner:
        description: Owner holds the value of the owner edge.
        items:
          $ref: '#/definitions/users-service_app_models_ent.User'
        type: array
      permissions:
        description: Permissions holds the value of the permissions edge.
        items:
          $ref: '#/definitions/users-service_app_models_ent.RolePermission'
        type: array
    type: object
  users-service_app_models_ent.UserSettings:
    properties:
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.UserSettingsEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserSettingsQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      night_mode:
        description: NightMode holds the value of the "night_mode" field.
        type: boolean
      two_factor:
        description: TwoFactor holds the value of the "two_factor" field.
        type: boolean
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      uuid:
        description: UUID holds the value of the "UUID" field.
        type: string
    type: object
  users-service_app_models_ent.UserSettingsEdges:
    properties:
      notification_target_devices:
        description: NotificationTargetDevices holds the value of the notification_target_devices
          edge.
        items:
          $ref: '#/definitions/users-service_app_models_ent.UserDevice'
        type: array
      owner:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.User'
        description: Owner holds the value of the owner edge.
      second_factor_target:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.UserDevice'
        description: SecondFactorTarget holds the value of the second_factor_target
          edge.
    type: object
  users-service_app_models_ent.VerificationCode:
    properties:
      code:
        description: Code holds the value of the "code" field.
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.VerificationCodeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the VerificationCodeQuery when eager-loading is set.
      expires_at:
        description: ExpiresAt holds the value of the "expires_at" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
    type: object
  users-service_app_models_ent.VerificationCodeEdges:
    properties:
      owner:
        allOf:
        - $ref: '#/definitions/users-service_app_models_ent.User'
        description: Owner holds the value of the owner edge.
    type: object
...
  users-service_app_models_requests.Details:
    type: object
...
  users-service_app_models_requests.Settings:
    properties:
      nightMode:
        type: boolean
      notificationsTargetDeviceIDs:
        items:
          type: integer
        type: array
      secondFactorTargetID:
        type: integer
      twoFactor:
        type: boolean
    type: object
  users-service_app_models_requests.User:
    properties:
      mail:
        type: string
      phone:
        type: string
    type: object
  users-service_app_models_responses.ErrorResponse:
    properties:
      error:
        type: string
      requestID:
        type: string
      status:
        type: string
    type: object
  users-service_app_models_responses.SuccessResponse:
    properties:
      data: {}
      requestID:
        type: string
      status:
        type: string
    type: object
  users-service_app_models_responses.User:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/users-service_app_models_ent.User'
    type: object
info:
  contact: {}
paths:
...
  /api/v1/users/details/update:
    patch:
      consumes:
      - application/json
      description: Updates the authenticated user's additional details.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User details to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users-service_app_models_requests.Details'
      produces:
      - application/json
      responses:
        "200":
          description: OK - User details updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/users-service_app_models_responses.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/users-service_app_models_responses.User'
              type: object
        "400":
          description: Bad Request - Invalid input data
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
        "422":
          description: Unprocessable Entity - Update failed
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
      summary: Update user's details
      tags:
      - users
  /api/v1/users/remove:
    delete:
      consumes:
      - application/json
      description: Permanently removes the authenticated user's account. This action
        is irreversible.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - Account removed successfully
          schema:
            allOf:
            - $ref: '#/definitions/users-service_app_models_responses.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
        "422":
          description: Unprocessable Entity - Removal failed
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
      summary: Remove user account
      tags:
      - users
  /api/v1/users/settings/update:
    patch:
      consumes:
      - application/json
      description: Updates the authenticated user's application settings (e.g., 2FA,
        night mode).
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User settings to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users-service_app_models_requests.Settings'
      produces:
      - application/json
      responses:
        "200":
          description: OK - User settings updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/users-service_app_models_responses.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/users-service_app_models_responses.User'
              type: object
        "400":
          description: Bad Request - Invalid input data
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
        "422":
          description: Unprocessable Entity - Update failed
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
      summary: Update user's settings
      tags:
      - users
  /api/v1/users/update:
    patch:
      consumes:
      - application/json
      description: Updates the authenticated user's basic profile information (e.g.,
        name, email).
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User data to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users-service_app_models_requests.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK - User updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/users-service_app_models_responses.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/users-service_app_models_responses.User'
              type: object
        "400":
          description: Bad Request - Invalid input data
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
        "422":
          description: Unprocessable Entity - Update failed
          schema:
            $ref: '#/definitions/users-service_app_models_responses.ErrorResponse'
      summary: Update user's basic information
      tags:
      - users
swagger: "2.0"